hw #11/RemoteDrawing/requirements.txt
flask
jsonpickle
opencv-python
numpy
pillow
